TLCPM-309: update build for lowercase gem name. restructure tests.
