+++ test/test_helper.rb
@@ -1,9 +1,5 @@
 ## Common test setup / utilities
 
-
-
-#require_relative '../groups_admin'
-
 require 'minitest'
 require 'minitest/autorun'
 require 'minitest/unit'
@@ -14,8 +10,8 @@ require 'shoulda'
 require 'rack/test'
 
 TEST_DOMAIN='discussions-dev.its.umich.edu'
+EMAIL_INSERT_TEST_GROUP="ggb-test-group-insert@discussions-dev.its.umich.edu"
 
-#class TestHelper
 module TestHelper
 
   #include Rack::Test::Methods
@@ -45,7 +41,26 @@ module TestHelper
 
   # see if status has common messages
   def check_error_response(last_response)
-    refute_match 'Sinatra doesn&rsquo;t know this ditty',last_response.body,"unmatched url"
+    refute_match 'Sinatra doesn&rsquo;t know this ditty', last_response.body, "unmatched url"
   end
+
+
+  #From: "Alice Smith" <alice@example.com>
+  def create_test_email(group_id, from_name, from_email)
+    # Format an RFC822 message
+    now = Time.now
+    message_id = "#{now.to_f}-#{group_id}"
+    message_date = now.strftime '%a, %d %b %Y %T %z'
+    message = <<-EOF
+Message-ID: <#{message_id}>
+Date: #{message_date}
+To: #{group_id}
+From: "#{from_name}" <#{from_email}>
+Subject: Groups Migration API Test #{now.iso8601}
+
+This is a test email generated at #{now.iso8601}
+    EOF
+  end
+
 
 end
